impl method

+++ packages/mirror_sync.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 from packages.directory_obj import DirectoryObject
 from shutil import copy2, move
+from pprint import pprint
 
 
 class MirrorSync:
@@ -13,25 +14,63 @@ class MirrorSync:
         self.dest = DirectoryObject(destination) # DirectoryObjectを合成
 
 
-    def sync_exec(self, *, delete_ok=False):
+    def sync_exec(self, *, remove_ok=False):
         '''
         sourceとdestinationをシンクさせる
         delete_ok=Trueで、dest内のsrcに無いファイルと重複を削除する
         この順番で実行しないと予期しない挙動となる恐れがあります
         '''
-        if delete_ok == True:
-            self._delete_duplicate()
+        self._info_duplicate()
+        self._remove_duplicate(remove_ok=remove_ok)
         self._make_dir()
         self._copy_file()
-        if delete_ok == True:
-            self._remove_items()
+        self._remove_items(remove_ok=remove_ok)
         self._move_files()
 
+        self.src = DirectoryObject(self.src.base) # DirectoryObjectを再生成
+        self.dest = DirectoryObject(self.dest.base) # DirectoryObjectを再生成
 
-    def _delete_duplicate(self):
+
+    def _remove_duplicate(self, *, remove_ok=False):
         '''
         dest内に重複があれば削除する
         '''
+        dest_files = set()
+        to_remove = []
+
+        for file in self.dest.files:
+            if file.stem in dest_files:
+                to_remove.append(file)
+            else:
+                dest_files.add(file.stem)
+        
+        if remove_ok == True:
+            for p in to_remove:
+                # 削除処理
+                p.unlink()
+                self.dest.files.remove(p)
+                print(f'[removed] {str(p)}')
+        else:
+            for p in to_remove:
+                print(f'[duplicated] {str(p.stem)}')
+
+
+    def _info_duplicate(self):
+        '''
+        src内に重複があれば知らせる
+        '''
+        src_files = set()
+        to_removes = []
+
+        for file in self.src.files:
+            if file.stem in src_files:
+                to_removes.append(file)
+            else:
+                src_files.add(file.stem)
+        
+        for p in to_removes:
+            print(f'[duplicated] {str(p.stem)}')
+
 
     def _make_dir(self):
         '''
@@ -57,27 +96,35 @@ class MirrorSync:
                 print(f'[copied] {str(p)}')
             
 
-    def _remove_items(self):
+    def _remove_items(self, *, remove_ok=False):
         '''
         srcに無いfileとdirの削除
         '''
         # fileの削除
         source_files_stem = set(file.stem for file in self.src.files)
-        files_to_delete = []
+        to_removes = []
         for p in self.dest.files:
             if p.stem not in source_files_stem:
-                files_to_delete.append(p)
-        for p in files_to_delete:
-                p.unlink()
-                self.dest.files.remove(p)
-                print(f'[deleted] {str(p)}')
+                to_removes.append(p)
+        for p in to_removes:
+                if remove_ok == True:
+                    # 削除処理
+                    p.unlink()
+                    self.dest.files.remove(p)
+                    print(f'[removed] {str(p)}')
+                else:
+                    print(f'[not in source] {str(p)}')
         # dirの削除
         source_dirs = set(self.__substitute(dir) for dir in self.src.dirs)
         for p in self.dest.dirs:
             if p not in source_dirs:
-                p.rmdir()
-                self.dest.dirs.remove(p)
-                print(f'[deleted] {str(p)}')
+                if remove_ok == True:
+                    # 削除処理
+                    p.rmdir()
+                    self.dest.dirs.remove(p)
+                    print(f'[removed] {str(p)}')
+                else:
+                    print(f'[not in source] {str(p)}')
 
 
     def _move_files(self):
